fix: different align request in LMA VMA, affect data load
[Reason]:
For longlong test, when VMA in psram, LMA in flash for .text, .rodata and .data, header of .data in VMA is (at least) 8 bytes aligned, while LMA is not. However, this will not happen when both VMA and LMA in flash.
I guess align setting of VMA was set to 8 because there is longlong type in the test. But when both VMA and LMA are in psram, setting of LMA is 8 too, I have no idea on this.
* * * * * * * * * * *
[Fix]:
Set LMA and VMA of .data 16 bytes aligned explicitly. OR YOU CAN LOAD THE SEGMENTS ONE BY ONE.